// template/react-starter-kit/src/.generated/api/model/supportTicketPriority.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type SupportTicketPriority = typeof SupportTicketPriority[keyof typeof SupportTicketPriority];


export const SupportTicketPriority = {
  low: 'low',
  normal: 'normal',
  high: 'high',
  urgent: 'urgent',
} as const;