// template/react-starter-kit/src/.generated/api/model/supportTicketStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type SupportTicketStatus = typeof SupportTicketStatus[keyof typeof SupportTicketStatus];


export const SupportTicketStatus = {
  open: 'open',
  in_progress: 'in_progress',
  resolved: 'resolved',
  closed: 'closed',
} as const;